+++ src/Neutrals.jl
@@ -33,8 +33,8 @@ const Operand{T<:Number} = Union{T,AbstractArray{<:T}}
     Neutral{V}()
     Neutral(V)
 
-yield a neutral number of value `V` (one of `0`, `1`, or `-1`) which implement the
-following properties:
+yield a neutral number of value `V` (one of `0`, `1`, or `-1`) with the following
+properties:
 
 * `Neutral(0)` is the neutral element for the addition of dimensionless numbers and is a
   *strong zero* for the multiplication of numbers.
@@ -60,29 +60,26 @@ struct Neutral{V} <: Integer
     Neutral{-1}() = new{-1}() # Neutral(-1) is needed for arithmetic.
 end
 
-"""
-    Neutrals.value(x)
-    Neutrals.value(typeof(x))
-
-yield the value associated with neutral number `x`.
-
-"""
-value(::Neutral{x}) where x = x
-value(::Type{Neutral{x}}) where x = x
-
 # For numbers, there is no needs to extend `Base.convert` with the following "conversion"
 # constructors as `Base.convert(T, x)` falls back to call `T(x)::T`.
 Neutral(x::Neutral) = x
-Neutral(x::Number) = Neutral(Int(x)::Int)
-Neutral(x::Int) = Neutral{x}()
 Neutral{V}(x::Neutral{V}) where {V} = x
 Neutral{V}(x::Neutral) where {V} = throw(InexactError(:convert, Neutral{V}, x))
-Neutral{V}(x::Number) where {V} = x == V ? Neutral{V}() : throw(InexactError(:convert, Neutral{V}, x))
+Neutral(x::Int) = Neutral{x}()
+for type in (:Number, :Rational, :Complex, :BigFloat) # `Number` is not enough to get rid of ambiguities
+    @eval begin
+        Neutral(x::$type) = Neutral(Int(x)::Int)
+        Neutral{V}(x::$type) where {V} =
+            x == V ? Neutral{V}() : throw(InexactError(:convert, Neutral{V}, x))
+    end
+end
 
 # Error catcher.
 Neutral{V}() where {V} = throw(ArgumentError(
     "value "*(V isa Number ? repr(V) : "of type `$(typeof(V))`")
     *" cannot be converted into a neutral number"))
+
+@eval Base.instances(::Type{Neutral}) = $(map(Neutral, (0, 1, -1)))
 
 """
     ZERO
@@ -162,8 +159,6 @@ if VERSION ≥ v"1.3"
     eval(Meta.parse("export 𝟙; const 𝟙 = ONE; @doc (@doc ONE) 𝟙;"))
 end
 
-@eval Base.instances(::Type{Neutral}) = $(map(Neutral, (0, 1, -1)))
-
 # Aliases to encode rules. `AnyNeutral` is more specific than `Neutral` and is a union of
 # concrete types.
 const AnyNeutral = Union{map(typeof, instances(Neutral))...}
@@ -181,9 +176,9 @@ const UnsignedNumber = Union{UnsignedReal,UnsignedComplex}
 const BareNumber = Union{Real,Complex}
 
 Base.typemin(::Type{Neutral}) = -ONE
-Base.typemin(::Type{Neutral{x}}) where {x} = Neutral{x}()
+Base.typemin(::Type{<:Neutral{x}}) where {x} = Neutral{x}()
 Base.typemax(::Type{Neutral}) = ONE
-Base.typemax(::Type{Neutral{x}}) where {x} = Neutral{x}()
+Base.typemax(::Type{<:Neutral{x}}) where {x} = Neutral{x}()
 
 TypeUtils.is_signed(::Type{<:Neutral}) = true
 
@@ -202,6 +197,16 @@ for (T, name, descr) in ((Neutral{0}, "𝟘",
 end
 
 #---------------------------------------------------------------------------- CONVERSION -
+
+"""
+    Neutrals.value(x)
+    Neutrals.value(typeof(x))
+
+yield the value associated with neutral number `x`.
+
+"""
+value(::Neutral{x}) where x = x
+value(::Type{<:Neutral{x}}) where x = x
 
 # Conversion rules for bare numeric types. No needs to extend `Base.convert` because
 # `Base.convert(T,x)` amounts to calling `T(x)` for any numeric type `T`.
@@ -346,8 +351,8 @@ Base.promote_rule(::Type{<:Neutral}, ::Type{<:Neutral}) = Int
 Base.promote_rule(::Type{<:Neutral}, ::Type{T}) where {T<:Number} = T
 Base.promote_rule(::Type{<:Neutral}, ::Type{<:AbstractIrrational}) = Float64
 Base.promote_rule(::Type{Bool}, ::Type{<:Neutral}) = Bool
-Base.promote_rule(::Type{Bool}, ::Type{Neutral{-1}}) = Int
-Base.promote_rule(::Type{Neutral{-1}}, ::Type{Bool}) = Int
+Base.promote_rule(::Type{Bool}, ::Type{<:Neutral{-1}}) = Int
+Base.promote_rule(::Type{<:Neutral{-1}}, ::Type{Bool}) = Int
 
 #--------------------------------------------------------------------- BINARY OPERATIONS -
 
@@ -382,17 +387,18 @@ Base.promote_rule(::Type{Neutral{-1}}, ::Type{Bool}) = Int
 # and so on.
 #
 # Override base methods to call corresponding implementation:
-for (f, (g, w, Ts)) in (:(+)   => (:impl_add,    3, (:Number, :Real, :Integer, :Rational,
+for (f, (g, w, Ts)) in (:(+)   => (:impl_add,    6, (:Number, :Real, :Integer, :Rational,
                                                      :Complex, :(Complex{Bool}), :AbstractIrrational,
                                                      :AbstractFloat, :BigInt, :BigFloat)),
                         :(-)   => (:impl_sub,    3, (:Number, :(Complex{Bool}), :Real, :Integer, :Rational,
                                                      :Complex, :AbstractIrrational,
                                                      :AbstractFloat, :BigInt, :BigFloat)),
-                        :(*)   => (:impl_mul,    3, (:Number, :Real, :Integer,
+                        :(*)   => (:impl_mul,    5, (:Number, :Real, :Integer,
                                                      :Rational, :Complex, :(Complex{Bool}))),
                         :(/)   => (:impl_div,    3, (:Number, :Real, :Integer, :Rational,
                                                      :Complex, :(Complex{Bool}))),
                         :(^)   => (:impl_pow,    2, (:Number, :Real, :Integer, :Rational,
+                                                     :(Irrational{:ℯ}),
                                                      :Float16, :Float32, :Float64,
                                                      :Complex,
                                                      :(Complex{<:AbstractFloat}),
@@ -402,27 +408,38 @@ for (f, (g, w, Ts)) in (:(+)   => (:impl_add,    3, (:Number, :Real, :Integer, :
                         :div   => (:impl_tdv,    3, (:Real, :Rational)),
                         :rem   => (:impl_rem,    3, (:Real, :Rational)),
                         :mod   => (:impl_mod,    3, (:Real, :Rational)),
-                        :(==)  => (:impl_eq,     3, (:Number, :Real, :Rational,
+                        :(==)  => (:impl_eq,     6, (:Number, :Real, :Rational,
                                                      :AbstractIrrational, :Complex,
                                                      :BigInt, :BigFloat)),
                         :(<)   => (:impl_lt,     3, (:Real, :Rational, :BigInt, :BigFloat)),
                         :(<=)  => (:impl_le,     3, (:Real, :Rational, :BigInt, :BigFloat)),
                         :cmp   => (:impl_cmp,    3, (:Number, :Real, :Integer,
                                                      :BigInt, :BigFloat)),
-                        :(|)   => (:impl_or,     3, (:Integer,)),
-                        :(&)   => (:impl_and,    3, (:Integer,)),
-                        :xor   => (:impl_xor,    3, (:Integer,)),
+                        :(|)   => (:impl_or,     6, (:Integer,)),
+                        :(&)   => (:impl_and,    6, (:Integer,)),
+                        :xor   => (:impl_xor,    6, (:Integer,)),
                         :(<<)  => (:impl_lshft,  2, (:Integer, :Bool)),
                         :(>>)  => (:impl_rshft,  2, (:Integer, :Bool)),
                         :(>>>) => (:impl_urshft, 2, (:Integer, :Bool)),
                         )
-    @eval Base.$f(x::Neutral, y::Neutral) = $g(x, y)
+    # There is always an implementation when both operands are neutral numbers.
+    @eval Base.$f(x::Neutral, y::Neutral) = $g($(Val(3)), x, y)
     for T in Ts
-        if (w & 1) != 0
-            @eval Base.$f(x::Neutral, y::$T) = $g(x, y)
+        if (w & 1) == 1
+            # Implementation exists when 1st operand is a neutral number.
+            @eval Base.$f(x::Neutral, y::$T) = $g($(Val(1)), x, y)
+        elseif w == 6
+            # Operation is commutative and implementation exists when 2nd operand is a
+            # neutral number.
+            @eval Base.$f(x::Neutral, y::$T) = $g($(Val(2)), y, x)
         end
-        if (w & 2) != 0
-            @eval Base.$f(x::$T, y::Neutral) = $g(x, y)
+        if (w & 2) == 2
+            # Implementation exists when 2nd operand is a neutral number.
+            @eval Base.$f(x::$T, y::Neutral) = $g($(Val(2)), x, y)
+        elseif w == 5
+            # Operation is commutative and implementation exists when 1st operand is a
+            # neutral number.
+            @eval Base.$f(x::$T, y::Neutral) = $g($(Val(1)), y, x)
         end
     end
 end
@@ -445,11 +462,11 @@ for x in instances(Neutral), y in instances(Neutral)
                    :(>>>) => :impl_urshft)
         r = @eval $f(value($x), value($y))
         if r isa Bool
-            @eval $g(::$(typeof(x)), ::$(typeof(y))) = $r
+            @eval $g(::Val{3}, ::$(typeof(x)), ::$(typeof(y))) = $r
         elseif r ∈ (0, 1, -1) # returns a neutral number if possible
-            @eval $g(::$(typeof(x)), ::$(typeof(y))) = $(Neutral{r}())
+            @eval $g(::Val{3}, ::$(typeof(x)), ::$(typeof(y))) = $(Neutral{r}())
         else # otherwise returns an integer
-            @eval $g(::$(typeof(x)), ::$(typeof(y))) = $r
+            @eval $g(::Val{3}, ::$(typeof(x)), ::$(typeof(y))) = $r
         end
     end
 
@@ -459,22 +476,22 @@ for x in instances(Neutral), y in instances(Neutral)
                    :rem => :impl_rem,
                    :mod => :impl_mod)
         if y === ZERO
-            @eval $g(::$(typeof(x)), ::$(typeof(y))) = throw(DivideError())
+            @eval $g(::Val{3}, ::$(typeof(x)), ::$(typeof(y))) = throw(DivideError())
         else # y is ONE or -ONE
             r = if f === :(/)
                 value(x)*value(y) # x/y yields the same result as x*y when abs(y) = 1
             else
                 @eval $f(value($x), value($y))
             end
-            @eval $g(::$(typeof(x)), ::$(typeof(y))) = $(Neutral{r}())
+            @eval $g(::Val{3}, ::$(typeof(x)), ::$(typeof(y))) = $(Neutral{r}())
         end
     end
 
     # Exponentiation.
-    @eval impl_pow(::$(typeof(x)), ::$(typeof(y))) = $(y === ZERO ? ONE : x)
+    @eval impl_pow(::Val{3}, ::$(typeof(x)), ::$(typeof(y))) = $(y === ZERO ? ONE : x)
 
     # Comparison.
-    @eval impl_cmp(::$(typeof(x)), ::$(typeof(y))) = $(cmp(value(x), value(y)))
+    @eval impl_cmp(::Val{3}, ::$(typeof(x)), ::$(typeof(y))) = $(cmp(value(x), value(y)))
 end
 
 """
@@ -490,40 +507,41 @@ if VERSION < v"1.2.0-rc2"
 end
 
 """
-    Neutrals.impl_add(x, y) -> x + y
+    Neutrals.impl_add(w::Val, x, y) -> x + y
 
 implements addition of numbers `x` and `y` when at least one of the operands is a neutral
-number.
+number. To avoid ambiguities, the caller must set `w` to `Val(1)` if only `x` is a neutral
+number, `Val(2)` if only `y` is a neutral number, and `Val(3)` if `x` and `y` are both
+neutral numbers.
 
 This method can be overridden by specializing it when the second operand is a neutral
-number.
+number, that is for `w::Val{2}`.
 
 """
-impl_add(x::Neutral, y::Number) = impl_add(y, x) # put neutral number second
-impl_add(x::BareNumber, ::Neutral{0}) = x
-impl_add(x::Number, ::Neutral{ 0}) = is_dimensionless(x) ? x : throw_add_dimensionful_and_zero()
-impl_add(x::Number, ::Neutral{ 1}) = x + convert(type_common(x), 1)
-impl_add(x::Number, ::Neutral{-1}) = x - convert(type_common(x), 1)
+impl_add(::Val{2}, x::BareNumber, ::Neutral{0}) = x
+impl_add(::Val{2}, x::Number, ::Neutral{ 0}) = is_dimensionless(x) ? x : throw_add_dimensionful_and_zero()
+impl_add(::Val{2}, x::Number, ::Neutral{ 1}) = x + convert(type_common(x), 1)
+impl_add(::Val{2}, x::Number, ::Neutral{-1}) = x - convert(type_common(x), 1)
 
 @noinline throw_add_dimensionful_and_zero() =
     throw(ArgumentError("𝟘 and dimensionful quantity cannot be added"))
 
 """
-    Neutrals.impl_sub(x, y) -> x - y
+    Neutrals.impl_sub(w::Val, x, y) -> x - y
 
 implements subtraction of numbers `x` and `y` when at least one of the operands is a
-neutral number.
+neutral number. See [`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 """
-impl_sub(x::BareNumber, ::Neutral{0}) = x
-impl_sub(x::Number, ::Neutral{ 0}) = is_dimensionless(x) ? x : throw_sub_dimensionful_and_zero()
-impl_sub(x::Number, ::Neutral{ 1}) = x - convert(type_common(x), 1)
-impl_sub(x::Number, ::Neutral{-1}) = x + convert(type_common(x), 1)
+impl_sub(::Val{1}, x::Neutral{ 0}, y::BareNumber) = -y
+impl_sub(::Val{1}, x::Neutral{ 0}, y::Number) = is_dimensionless(y) ? -y : throw_sub_dimensionful_and_zero()
+impl_sub(::Val{1}, x::Neutral{ 1}, y::Number) = convert(type_common(y), 1) - y
+impl_sub(::Val{1}, x::Neutral{-1}, y::Number) = -convert(type_common(y), 1) - y
 
-impl_sub(::Neutral{ 0}, x::BareNumber) = -x
-impl_sub(::Neutral{ 0}, x::Number) = is_dimensionless(x) ? -x : throw_sub_dimensionful_and_zero()
-impl_sub(::Neutral{ 1}, x::Number) = convert(type_common(x), 1) - x
-impl_sub(::Neutral{-1}, x::Number) = -convert(type_common(x), 1) - x
+impl_sub(::Val{2}, x::BareNumber, y::Neutral{0}) = x
+impl_sub(::Val{2}, x::Number, y::Neutral{ 0}) = is_dimensionless(x) ? x : throw_sub_dimensionful_and_zero()
+impl_sub(::Val{2}, x::Number, y::Neutral{ 1}) = x - convert(type_common(x), 1)
+impl_sub(::Val{2}, x::Number, y::Neutral{-1}) = x + convert(type_common(x), 1)
 
 @noinline throw_sub_dimensionful_and_zero() =
     throw(ArgumentError("𝟘 and dimensionful quantity cannot be subtracted"))
@@ -531,76 +549,71 @@ impl_sub(::Neutral{-1}, x::Number) = -convert(type_common(x), 1) - x
 # In Julia, Booleans are promoted to `Int` for addition, subtraction and bitwise shifts
 # (base/bool.jl). The implementations of addition and subtraction of a Boolean with `±𝟙`
 # are specialized according to this.
-impl_add(x::Bool, y::Neutral{ 1}) = Int(x) + 1
-impl_add(x::Bool, y::Neutral{-1}) = Int(x) - 1
+impl_add(::Val{2}, x::Bool, y::Neutral{ 1}) = Int(x) + 1
+impl_add(::Val{2}, x::Bool, y::Neutral{-1}) = Int(x) - 1
 #
-impl_sub(x::Bool, ::Neutral{ 1}) = Int(x) - 1
-impl_sub(x::Bool, ::Neutral{-1}) = Int(x) + 1
-impl_sub(::Neutral{ 1}, x::Bool) = 1 - Int(x)
-impl_sub(::Neutral{-1}, x::Bool) = -1 - Int(x)
+impl_sub(::Val{2}, x::Bool, y::Neutral{ 1}) = Int(x) - 1
+impl_sub(::Val{2}, x::Bool, y::Neutral{-1}) = Int(x) + 1
+impl_sub(::Val{1}, x::Neutral{ 1}, y::Bool) = 1 - Int(y)
+impl_sub(::Val{1}, x::Neutral{-1}, y::Bool) = -1 - Int(y)
 
 """
-    Neutrals.impl_mul(x, y) -> x * y    # if x and y are numbers
-    Neutrals.impl_mul(x, y) -> x .* y   # if one of x or y is an array
+    Neutrals.impl_mul(w::Val, x, y) -> x * y    # if x and y are numbers
+    Neutrals.impl_mul(w::Val, x, y) -> x .* y   # if one of x or y is an array
 
 implement scalar or element-wise multiplication of `x` by `y` when at least one of the
-operands is a neutral number while the other is a number or an array of numbers.
+operands is a neutral number while the other is a number or an array of numbers. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 This method can be overridden by specializing it when the first operand is a neutral
-number.
+number, that is for `w::Val{1}`.
 
 """
-impl_mul(x::Operand{Number}, y::Neutral) = impl_mul(y, x) # put neutral number first
-
-# Scalar or element-wise multiplication `x*y` with `x` a neutral number.
-impl_mul(x::Neutral{ 0}, y::BareNumber) = ZERO
-impl_mul(x::Neutral{ 0}, y::AbstractArray{<:BareNumber}) = similar(y, Neutral{0})
-impl_mul(x::Neutral{ 1}, y::Operand{Number}) = y
-impl_mul(x::Neutral{-1}, y::Operand{Number}) = -y
-
-Base.:*(x::Neutral, y::AbstractArray{<:Number}) = impl_mul(x, y)
-Base.:*(x::AbstractArray{<:Number}, y::Neutral) = impl_mul(y, x) # put neutral number first
+impl_mul(::Val{1}, x::Neutral{ 0}, y::BareNumber) = ZERO
+impl_mul(::Val{1}, x::Neutral{ 0}, y::AbstractArray{<:BareNumber}) = similar(y, typeof(x))
+impl_mul(::Val{1}, x::Neutral{ 1}, y::Operand{Number}) = y
+impl_mul(::Val{1}, x::Neutral{-1}, y::Operand{Number}) = -y
 
 """
-    Neutrals.impl_div(x, y) -> x / y    # if x and y are numbers
-    Neutrals.impl_div(x, y) -> x ./ y   # if one of x or y is an array
+    Neutrals.impl_div(w::Val, x, y) -> x / y    # if x and y are numbers
+    Neutrals.impl_div(w::Val, x, y) -> x ./ y   # if one of x or y is an array
 
 implement scalar or element-wise division of `x` by `y` when at least one of the operands
-is a neutral number while the other is a number or an array of numbers.
+is a neutral number while the other is a number or an array of numbers. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 """
-impl_div(x::Operand{Number}, y::Neutral{0}) = throw(DivideError())
-impl_div(x::Operand{Number}, y::Neutral{1}) = x
-impl_div(x::Operand{Number}, y::Neutral{-1}) = -x
+impl_div(::Val{1}, x::Neutral{ 0}, y::BareNumber) = ZERO
+impl_div(::Val{1}, x::Neutral{ 1}, y::Number) = impl_inv(y)
+impl_div(::Val{1}, x::Neutral{-1}, y::Number) = -impl_inv(y)
 
-impl_div(x::Neutral{ 0}, y::BareNumber) = ZERO
-impl_div(x::Neutral{ 1}, y::Number) = impl_inv(y)
-impl_div(x::Neutral{-1}, y::Number) = -impl_inv(y)
+impl_div(::Val{2}, x::Operand{Number}, y::Neutral{0}) = throw(DivideError())
+impl_div(::Val{2}, x::Operand{Number}, y::Neutral{1}) = x
+impl_div(::Val{2}, x::Operand{Number}, y::Neutral{-1}) = -x
 
 # Element-wise division of neutral number `x` by array `y`. Division by zero is first
 # checked, then the result is computed according to the specific value of `x` using
 # auxiliary function `_impl_div` to dispatch on `x`.
-impl_div(x::Neutral, y::AbstractArray{Neutral{0}}) = throw(DivideError())
-impl_div(x::Neutral, y::AbstractArray{<:Number}) = _impl_div(x, y) # to dispatch on x
+impl_div(::Val{1}, x::Neutral, y::AbstractArray{<:Neutral{0}}) = throw(DivideError())
+impl_div(::Val{1}, x::Neutral, y::AbstractArray{<:Number}) = _impl_div(x, y) # to dispatch on x
 _impl_div(x::Neutral{ 0}, y::AbstractArray{<:BareNumber}) = similar(y, typeof(x))
 _impl_div(x::Neutral{ 1}, y::AbstractArray{<:Number}) = impl_inv.(y)
 _impl_div(x::Neutral{-1}, y::AbstractArray{<:Number}) = -impl_inv.(y)
 
-Base.:/(x::AbstractArray{<:Number}, y::Neutral) = impl_div(x, y)
-Base.:\(x::Neutral, y::AbstractArray{<:Number}) = impl_div(y, x)
-
 """
-    Neutrals.impl_tdv(x, y) -> x ÷ y
+    Neutrals.impl_tdv(w::Val, x, y) -> x ÷ y
 
 implements the truncated division of `x` by `y` that is the quotient of `x` by `y` rounded
-toward zero. This corresponds to `div(x,y,RoundToZero)`.
+toward zero. This corresponds to `div(x,y,RoundToZero)`. See [`Neutrals.impl_add`](@ref)
+for the interpretation of `w`.
 
 """ impl_tdv
 
 """
-    Neutrals.impl_rem(x, y) -> rem(x, y)
+    Neutrals.impl_rem(w::Val, x, y) -> rem(x, y)
 
 implements the remainder function when at least one of the operands is a neutral number.
+See [`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 The remainder function satisfies `x == div(x,y)*y + rem(x,y)` with `div` the truncated
 division which yields the quotient rounded toward zero, implying that sign matches `x`.
@@ -609,9 +622,10 @@ division which yields the quotient rounded toward zero, implying that sign match
 impl_rem
 
 """
-    Neutrals.impl_mod(x, y) -> mod(x, y)
+    Neutrals.impl_mod(w::Val, x, y) -> mod(x, y)
 
-implements `mod` when at least one of the operands is a neutral number.
+implements `mod` when at least one of the operands is a neutral number. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 The modulus function satisfies `x == fld(x,y)*y + mod(x,y)`, with `fld` the floored
 division which yields the rounded towards `-Inf`, implying that sign matches `y`.
@@ -624,25 +638,25 @@ division which yields the rounded towards `-Inf`, implying that sign matches `y`
 # other operand.
 for (f, g) in (:div => :impl_tdv, :rem => :impl_rem, :mod => :impl_mod)
     @eval begin
-        $g(x::Real, y::Neutral{0}) = throw(DivideError())
-        $g(x::Real, y::Neutral) = $f(x, convert(type_signed(x), y))
-        $g(x::Neutral, y::Real) = $f(convert(type_signed(y), x), y)
+        $g(::Val{2}, x::Real, y::Neutral{0}) = throw(DivideError())
+        $g(::Val{2}, x::Real, y::Neutral) = $f(x, convert(type_signed(x), y))
+        $g(::Val{1}, x::Neutral, y::Real) = $f(convert(type_signed(y), x), y)
     end
 end
 
-# Specialize implementation for integers(not Booleans) and for `±𝟙` considering that
+# Specialize implementation for integers (not Booleans) and for `±𝟙` considering that
 # neutral numbers are signed integers.
 #
 # Quotient of truncated division is of the signedness of the 1st operand.
-impl_tdv(x::Integer, y::Neutral{1}) = x
+impl_tdv(::Val{2}, x::Integer, y::Neutral{1}) = x
 #
 # Remainder of truncated division is of the signedness of the `st operand.
-impl_rem(x::Integer, y::Neutral{1}) = zero(x) # FIXME yield ZERO instead?
-impl_rem(x::Signed, y::Neutral{-1}) = zero(x) # FIXME yield ZERO instead?
+impl_rem(::Val{2}, x::Integer, y::Neutral{1}) = zero(x) # FIXME yield ZERO instead?
+impl_rem(::Val{2}, x::Signed, y::Neutral{-1}) = zero(x) # FIXME yield ZERO instead?
 #
 # Modulo is of the signedness of the 2nd operand and is 0 if 2nd operand is -1.
-impl_mod(x::Integer, y::Neutral{1}) = zero(type_signed(x)) # FIXME yield ZERO instead?
-impl_mod(x::Integer, y::Neutral{-1}) = zero(type_signed(x)) # FIXME yield ZERO instead?
+impl_mod(::Val{2}, x::Integer, y::Neutral{1}) = zero(type_signed(x)) # FIXME yield ZERO instead?
+impl_mod(::Val{2}, x::Integer, y::Neutral{-1}) = zero(type_signed(x)) # FIXME yield ZERO instead?
 #
 # For Booleans, implementation of `div`, `rem`, and `mod` in `base/bool.jl` is:
 #
@@ -650,124 +664,134 @@ impl_mod(x::Integer, y::Neutral{-1}) = zero(type_signed(x)) # FIXME yield ZERO i
 #     rem(x::Bool, y::Bool) = y ? false : throw(DivideError())
 #     mod(x::Bool, y::Bool) = rem(x,y)
 #
-impl_tdv(x::Bool, y::Neutral{1}) = x
-impl_rem(x::Bool, y::Neutral{1}) = false
-impl_mod(x::Bool, y::Neutral{1}) = false
+impl_tdv(::Val{2}, x::Bool, y::Neutral{1}) = x
+impl_rem(::Val{2}, x::Bool, y::Neutral{1}) = false
+impl_mod(::Val{2}, x::Bool, y::Neutral{1}) = false
 for f in (:impl_tdv, :impl_rem, :impl_mod)
-    @eval $f(x::Bool, y::Neutral{-1}) = throw(InexactError(:convert, Bool, -ONE))
+    @eval $f(::Val{2}, x::Bool, y::Neutral{-1}) = throw(InexactError(:convert, Bool, -ONE))
 end
 
 """
-    Neutrals.impl_pow(x, y) -> x^y
+    Neutrals.impl_pow(w::Val, x, y) -> x^y
 
-implements raising number `x` to the power `y` when `y` is a neutral number.
+implements raising number `x` to the power `y` when `y` is a neutral number. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 """
-impl_pow(x::Number, ::Neutral{0}) = oneunit(x)
-impl_pow(x::Number, ::Neutral{1}) = x
-impl_pow(x::Number, ::Neutral{-1}) = impl_inv(x)
+impl_pow(::Val{2}, x::Number, ::Neutral{0}) = oneunit(x)
+impl_pow(::Val{2}, x::Number, ::Neutral{1}) = x
+impl_pow(::Val{2}, x::Number, ::Neutral{-1}) = impl_inv(x)
 
 # There is no `oneunit` for irrational numbers.
-impl_pow(x::AbstractIrrational, ::Neutral{0}) = 1.0
+impl_pow(::Val{2}, x::AbstractIrrational, ::Neutral{0}) = 1.0
 
 """
-    Neutrals.impl_eq(x, y) -> x == y
+    Neutrals.impl_eq(w::Val, x, y) -> x == y
 
-implements `==` for numbers when at least one of the operands is a neutral number.
+implements `==` for numbers when at least one of the operands is a neutral number. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 This method can be overridden by specializing it when the second operand is a neutral
-number.
+number, that is for for `w::Val{2}`.
 
 """
-impl_eq(x::Neutral, y::Number) = impl_eq(y, x) # put neutral number second
-impl_eq(x::Number, ::Neutral{ 0}) = is_dimensionless(x) && iszero(x)
-impl_eq(x::Number, ::Neutral{ 1}) = isone(x)
-impl_eq(x::Number, ::Neutral{-1}) = x == -convert(type_common(x), 1)
+impl_eq(::Val{2}, x::Number, ::Neutral{ 0}) = is_dimensionless(x) && iszero(x)
+impl_eq(::Val{2}, x::Number, ::Neutral{ 1}) = isone(x)
+impl_eq(::Val{2}, x::Number, ::Neutral{-1}) = x == -convert(type_common(x), 1)
  # NOTE We are assuming that `isone(x)` is not slower than `x == one(x)` or `x ==
  #      convert(type_common(x), 1)`. We assume that `x == -𝟙` is specialized to yield
  #      `false` for unsigned numbers (see below).
 
 # Optimize comparison of an unsigned real and a neutral number.
-impl_eq(x::UnsignedNumber, y::Neutral{-1}) = false
+impl_eq(::Val{2}, x::UnsignedNumber, y::Neutral{-1}) = false
 #
-impl_eq(x::Bool, y::Neutral{1}) = x
-impl_eq(x::Bool, y::Neutral{0}) = !x
+impl_eq(::Val{2}, x::Bool, y::Neutral{1}) = x
+impl_eq(::Val{2}, x::Bool, y::Neutral{0}) = !x
 
 # Neutral numbers are integers and are never equal to irrational numbers.
 for n in instances(Neutral)
-    @eval impl_eq(x::AbstractIrrational, y::$(typeof(n))) = false
+    @eval impl_eq(::Val{2}, x::AbstractIrrational, y::$(typeof(n))) = false
 end
 
 """
-    Neutrals.impl_lt(x, y) -> x < y
+    Neutrals.impl_lt(w::Val, x, y) -> x < y
 
-implements `<` for real numbers when at least one of the operands is a neutral number.
+implements `<` for real numbers when at least one of the operands is a neutral number. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 """ impl_lt
 
 # Optimize comparison of an unsigned real and a neutral number.
-impl_lt(x::UnsignedReal, y::Neutral{0}) = false
-impl_lt(x::UnsignedReal, y::Neutral{-1}) = false
-impl_lt(x::Neutral{-1}, y::UnsignedReal) = true
+impl_lt(::Val{2}, x::UnsignedReal, y::Neutral{0}) = false
+impl_lt(::Val{2}, x::UnsignedReal, y::Neutral{-1}) = false
+impl_lt(::Val{1}, x::Neutral{-1}, y::UnsignedReal) = true
 #
-impl_lt(x::Bool, y::Neutral{1}) = !x
-impl_lt(x::Neutral{0}, y::Bool) = y
-impl_lt(x::Neutral{1}, y::Bool) = false
+impl_lt(::Val{2}, x::Bool, y::Neutral{1}) = !x
+impl_lt(::Val{1}, x::Neutral{0}, y::Bool) = y
+impl_lt(::Val{1}, x::Neutral{1}, y::Bool) = false
 
 """
-    Neutrals.impl_le(x, y) -> x ≤ y
+    Neutrals.impl_le(w::Val, x, y) -> x ≤ y
 
-implements `≤` for real numbers when at least one of the operands is a neutral number.
+implements `≤` for real numbers when at least one of the operands is a neutral number. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 """ impl_le
 
 # Optimize comparison of an unsigned real and a neutral number.
-impl_le(x::UnsignedReal, y::Neutral{-1}) = false
-impl_le(x::Neutral{0}, y::UnsignedReal) = true
-impl_le(x::Neutral{-1}, y::UnsignedReal) = true
+impl_le(::Val{2}, x::UnsignedReal, y::Neutral{-1}) = false
+impl_le(::Val{1}, x::Neutral{0}, y::UnsignedReal) = true
+impl_le(::Val{1}, x::Neutral{-1}, y::UnsignedReal) = true
 #
-impl_le(x::Bool, y::Neutral{0}) = !x
-impl_le(x::Bool, y::Neutral{1}) = true
-impl_le(x::Neutral{1}, y::Bool) = y
+impl_le(::Val{2}, x::Bool, y::Neutral{0}) = !x
+impl_le(::Val{2}, x::Bool, y::Neutral{1}) = true
+impl_le(::Val{1}, x::Neutral{1}, y::Bool) = y
 
 for (f, g) in (:(<) => :impl_lt, :(<=) => :impl_le)
     @eval begin
-        $g(x::Real, y::Neutral) = $f(x, convert(type_common(x), y))
-        $g(x::Neutral, y::Real) = $f(convert(type_common(y), x), y)
+        $g(::Val{2}, x::Real, y::Neutral) = $f(x, convert(type_common(x), y))
+        $g(::Val{1}, x::Neutral, y::Real) = $f(convert(type_common(y), x), y)
     end
 end
 
 """
-    Neutrals.impl_cmp(x, y) -> cmp(x, y)
+    Neutrals.impl_cmp(w::Val, x, y) -> cmp(x, y)
 
 implements `cmp` for real numbers when at least one of the operands is a neutral number.
+See [`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 This method can be overridden by specializing it when the second operand is a neutral
 number.
 
 """
-impl_cmp(x::Neutral, y::Real) = -impl_cmp(y, x) # put neutral number second
-impl_cmp(x::Integer, y::Neutral) = ifelse(impl_isless(x, y), -1, ifelse(impl_isless(y, x), 1, 0))
-impl_cmp(x::Real, y::Neutral) = impl_isless(x, y) ? -1 : ifelse(impl_isless(y, x), 1, 0)
+impl_cmp(::Val{1}, x::Neutral, y::Real) =
+    -impl_cmp(Val(2), y, x) # put neutral number second
+impl_cmp(::Val{2}, x::Integer, y::Neutral) =
+    ifelse(impl_isless(Val(2), x, y), -1, ifelse(impl_isless(Val(1), y, x), 1, 0))
+impl_cmp(::Val{2}, x::Real, y::Neutral) =
+    impl_isless(Val(2), x, y) ? -1 : ifelse(impl_isless(Val(1), y, x), 1, 0)
 
 # Optimize comparison of an unsigned real and a neutral number.
-impl_cmp(x::UnsignedReal, y::Neutral{-1}) = 1
-impl_cmp(x::UnsignedReal, y::Neutral{0}) = iszero(x) ? 0 : 1
+impl_cmp(::Val{2}, x::UnsignedReal, y::Neutral{-1}) = 1
+impl_cmp(::Val{2}, x::UnsignedReal, y::Neutral{0}) = iszero(x) ? 0 : 1
 #
-impl_cmp(x::Bool, y::Neutral{1}) = x ? 0 : -1
+impl_cmp(::Val{2}, x::Bool, y::Neutral{1}) = x ? 0 : -1
 
 """
-    Neutrals.impl_isless(x, y) -> isless(x, y)
+    Neutrals.impl_isless(w::Val, x, y) -> isless(x, y)
 
-implements `isless` for real numbers when at least one of the operands is a neutral number.
+implements `isless` for real numbers when at least one of the operands is a neutral
+number. See [`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 """
-@inline impl_isless(x::Real, y::Real) = impl_lt(x, y)
+@inline impl_isless(w::Val, x::Real, y::Real) = impl_lt(w, x, y)
 
 # NOTE For floats in `base/float.jl`:
 #      isless(x, y) =  isnan(x) || isnan(b) ? !isnan(x) : x < y
-@inline impl_isless(x::AbstractFloat, y::Neutral) = isnan(x) ? false : x < oftype(x, value(y))
-@inline impl_isless(x::Neutral, y::AbstractFloat) = isnan(y) ? true : oftype(y, value(x)) < y
+@inline impl_isless(::Val{2}, x::AbstractFloat, y::Neutral) =
+    isnan(x) ? false : x < oftype(x, value(y))
+@inline impl_isless(::Val{1}, x::Neutral, y::AbstractFloat) =
+    isnan(y) ? true : oftype(y, value(x)) < y
 
 # For bitwise operations (`|`, `&`, and `xor`) between integers (including Booleans and
 # big integers) of mixed types, the called methods are defined in `base/int.jl` and
@@ -777,54 +801,54 @@ implements `isless` for real numbers when at least one of the operands is a neut
 # type with all its bits set to 1.
 
 """
-    Neutrals.impl_or(x, y)
+    Neutrals.impl_or(w::Val, x, y) -> x | y
 
 yields `x | y` when one of the operands is a neutral number while the other is an integer.
-If this method is overridden, it is sufficient to specialize it when the second operand is
-a neutral number.
+See [`Neutrals.impl_add`](@ref) for the interpretation of `w`. If this method is
+overridden, it is sufficient to specialize it when the second operand is a neutral number,
+that is for `w::Val{2}`.
 
 """
-impl_or(x::Neutral, y::Integer) = impl_or(y, x) # put neutral neutral second
-impl_or(x::Integer, ::Neutral{0}) = x
-impl_or(x::Integer, ::Neutral{1}) = x | one(x)
-impl_or(x::Integer, ::Neutral{-1}) = ~zero(x) # NOTE see remark for `x & 𝟘`
+impl_or(::Val{2}, x::Integer, ::Neutral{0}) = x
+impl_or(::Val{2}, x::Integer, ::Neutral{1}) = x | one(x)
+impl_or(::Val{2}, x::Integer, ::Neutral{-1}) = ~zero(x) # NOTE see remark for `x & 𝟘`
 
 # Optimize for Booleans.
-impl_or(x::Bool, ::Neutral{1}) = true
-impl_or(x::Bool, ::Neutral{-1}) = true
+impl_or(::Val{2}, x::Bool, ::Neutral{1}) = true
+impl_or(::Val{2}, x::Bool, ::Neutral{-1}) = true
 
 """
-    Neutrals.impl_and(x, y)
+    Neutrals.impl_and(w::Val, x, y) -> x & y
 
 yields `x & y` when one of the operands is a neutral number while the other is an integer.
-If this method is overridden, it is sufficient to specialize it when the second operand is
-a neutral number.
+See [`Neutrals.impl_add`](@ref) for the interpretation of `w`. If this method is
+overridden, it is sufficient to specialize it when the second operand is a neutral number,
+that is for `w::Val{2}`.
 
 """
-impl_and(x::Neutral, y::Integer) = impl_and(y, x) # operation is commutative
-impl_and(x::Integer, ::Neutral{0}) = zero(x) # NOTE not 𝟘, because 𝟘 is defined according to + and *, not &
-impl_and(x::Integer, ::Neutral{1}) = x & one(x)
-impl_and(x::Integer, ::Neutral{-1}) = x
+impl_and(::Val{2}, x::Integer, ::Neutral{0}) = zero(x) # NOTE not 𝟘, because 𝟘 is defined according to + and *, not &
+impl_and(::Val{2}, x::Integer, ::Neutral{1}) = x & one(x)
+impl_and(::Val{2}, x::Integer, ::Neutral{-1}) = x
 
 # Optimize for Booleans.
-impl_and(x::Bool, ::Neutral{1}) = x
+impl_and(::Val{2}, x::Bool, ::Neutral{1}) = x
 
 """
-    Neutrals.impl_xor(x, y)
+    Neutrals.impl_xor(w::Val, x, y)
 
 yields `xor(x, y)` when one of the operands is a neutral number while the other is an
-integer. If this method is overridden, it is sufficient to specialize it when the second
-operand is a neutral number.
+integer. See [`Neutrals.impl_add`](@ref) for the interpretation of `w`. If this method is
+overridden, it is sufficient to specialize it when the second operand is a neutral number,
+that is for `w::Val{2}`.
 
 """
-impl_xor(x::Neutral, y::Integer) = impl_xor(y, x) # operation is commutative
-impl_xor(x::Integer, ::Neutral{0}) = x
-impl_xor(x::Integer, ::Neutral{1}) = xor(x, one(x))
-impl_xor(x::Integer, ::Neutral{-1}) = xor(x, ~zero(x))
+impl_xor(::Val{2}, x::Integer, y::Neutral{0}) = x
+impl_xor(::Val{2}, x::Integer, y::Neutral{1}) = xor(x, one(x))
+impl_xor(::Val{2}, x::Integer, y::Neutral{-1}) = xor(x, ~zero(x))
 
 # Optimize for Booleans.
-impl_xor(x::Bool, ::Neutral{1}) = !x
-impl_xor(x::Bool, ::Neutral{-1}) = !x
+impl_xor(::Val{2}, x::Bool, ::Neutral{1}) = !x
+impl_xor(::Val{2}, x::Bool, ::Neutral{-1}) = !x
 
 # For bit shift operation of an integer `x` (including Booleans and big integers) by a
 # number of bits specified by a neutral number, it is sufficient to override the
@@ -833,34 +857,37 @@ impl_xor(x::Bool, ::Neutral{-1}) = !x
 # right operation with a number of bits specified as an `UInt` (see base/int.jl).
 
 """
-    Neutrals.impl_lshft(x, y) -> x << y
+    Neutrals.impl_lshft(w::Val{2}, x, y) -> x << y
 
-yields left bit shift of integer `x` by neutral number `y`.
-
-"""
-impl_lshft(x::Integer, ::Neutral{ 0}) = x
-impl_lshft(x::Integer, ::Neutral{ 1}) = x << UInt(1)
-impl_lshft(x::Integer, ::Neutral{-1}) = x >> UInt(1)
+yields left bit shift of integer `x` by neutral number `y`. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 """
-    Neutrals.impl_rshft(x, y) -> x >> y
-
-yields right bit shift of integer `x` by neutral number `y`.
-
-"""
-impl_rshft(x::Integer, ::Neutral{ 0}) = x
-impl_rshft(x::Integer, ::Neutral{ 1}) = x >> UInt(1)
-impl_rshft(x::Integer, ::Neutral{-1}) = x << UInt(1)
+impl_lshft(::Val{2}, x::Integer, ::Neutral{ 0}) = x
+impl_lshft(::Val{2}, x::Integer, ::Neutral{ 1}) = x << UInt(1)
+impl_lshft(::Val{2}, x::Integer, ::Neutral{-1}) = x >> UInt(1)
 
 """
-    Neutrals.impl_rshft(x, y) -> x >>> y
+    Neutrals.impl_rshft(w::Val{2}, x, y) -> x >> y
 
-yields unsigned right bit shift of integer `x` by neutral number `y`.
+yields right bit shift of integer `x` by neutral number `y`. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
 
 """
-impl_urshft(x::Integer, ::Neutral{ 0}) = x
-impl_urshft(x::Integer, ::Neutral{ 1}) = x >>> UInt(1)
-impl_urshft(x::Integer, ::Neutral{-1}) = x << UInt(1)
+impl_rshft(::Val{2}, x::Integer, ::Neutral{ 0}) = x
+impl_rshft(::Val{2}, x::Integer, ::Neutral{ 1}) = x >> UInt(1)
+impl_rshft(::Val{2}, x::Integer, ::Neutral{-1}) = x << UInt(1)
+
+"""
+    Neutrals.impl_rshft(w::Val{2}, x, y) -> x >>> y
+
+yields unsigned right bit shift of integer `x` by neutral number `y`. See
+[`Neutrals.impl_add`](@ref) for the interpretation of `w`.
+
+"""
+impl_urshft(::Val{2}, x::Integer, ::Neutral{ 0}) = x
+impl_urshft(::Val{2}, x::Integer, ::Neutral{ 1}) = x >>> UInt(1)
+impl_urshft(::Val{2}, x::Integer, ::Neutral{-1}) = x << UInt(1)
 
 """
     Neutrals.is_dimensionless(x)
@@ -886,9 +913,9 @@ is_dimensionless(::Type{<:BareNumber}) = true
 # larger than a `Clong` calls one of the compiled methods with `c` as a `Clong` or as a
 # `Culong`. Hence, we only have to specialize `cmp` for a big integer and a neutral
 # number.
-impl_cmp(x::BigInt, y::Neutral{ 0}) = cmp(x, Culong(0))
-impl_cmp(x::BigInt, y::Neutral{ 1}) = cmp(x, Culong(1))
-impl_cmp(x::BigInt, y::Neutral{-1}) = cmp(x, Clong(-1))
+impl_cmp(::Val{2}, x::BigInt, y::Neutral{ 0}) = cmp(x, Culong(0))
+impl_cmp(::Val{2}, x::BigInt, y::Neutral{ 1}) = cmp(x, Culong(1))
+impl_cmp(::Val{2}, x::BigInt, y::Neutral{-1}) = cmp(x, Clong(-1))
 #
 # As can be seen in `base/gmp.jl`, for the addition or subtraction of a big integer with
 # `c`, an integer of size ≤ `sizeof(Clong)`, the operation branches on the sign of `c` to
@@ -904,18 +931,18 @@ impl_cmp(x::BigInt, y::Neutral{-1}) = cmp(x, Clong(-1))
 for T in (:BigInt, :BigFloat)
     @eval begin
         # Addition. It is only needed to extend the rules for `±𝟙`.
-        impl_add(x::$T, y::Neutral{ 1}) = x + Culong(1)
-        impl_add(x::$T, y::Neutral{-1}) = x - Culong(1)
+        impl_add(::Val{2}, x::$T, y::Neutral{ 1}) = x + Culong(1)
+        impl_add(::Val{2}, x::$T, y::Neutral{-1}) = x - Culong(1)
 
         # Subtraction. It is only needed to extend the rules for `±𝟙`.
-        impl_sub(x::$T, y::Neutral{ 1}) = x - Culong(1)
-        impl_sub(x::$T, y::Neutral{-1}) = x + Culong(1)
+        impl_sub(::Val{2}, x::$T, y::Neutral{ 1}) = x - Culong(1)
+        impl_sub(::Val{2}, x::$T, y::Neutral{-1}) = x + Culong(1)
 
-        impl_sub(x::Neutral{ 1}, y::$T) = Culong(1) - y
-        impl_sub(x::Neutral{-1}, y::$T) = -(y + Culong(1))
+        impl_sub(::Val{1}, x::Neutral{ 1}, y::$T) = Culong(1) - y
+        impl_sub(::Val{1}, x::Neutral{-1}, y::$T) = -(y + Culong(1))
 
         # Equality. It is only needed to extend the rules for `-𝟙`.
-        impl_eq(x::$T, y::Neutral{-1}) = x == Clong(-1)
+        impl_eq(::Val{2}, x::$T, y::Neutral{-1}) = x == Clong(-1)
     end
 end
 
@@ -943,13 +970,13 @@ end
 
 broadcasted(::typeof(-), x::Operand{Number}, ::Neutral{0}) = x
 
-broadcasted(::typeof(*), x::Neutral, y::Neutral) = impl_mul(x, y)
-broadcasted(::typeof(*), x::Operand{Number}, y::Neutral) = impl_mul(y, x) # put neutral number 1st
-broadcasted(::typeof(*), x::Neutral, y::Operand{Number}) = impl_mul(x, y)
+broadcasted(::typeof(*), x::Neutral, y::Neutral) = impl_mul(Val(3), x, y)
+broadcasted(::typeof(*), x::Operand{Number}, y::Neutral) = impl_mul(Val(1), y, x) # put neutral number 1st
+broadcasted(::typeof(*), x::Neutral, y::Operand{Number}) = impl_mul(Val(1), x, y)
 
-broadcasted(::typeof(/), x::Neutral, y::Neutral) = impl_div(x, y)
-broadcasted(::typeof(/), x::Operand{Number}, y::Neutral) = impl_div(x, y)
-broadcasted(::typeof(/), x::Neutral, y::Operand{Number}) = impl_div(x, y)
+broadcasted(::typeof(/), x::Neutral, y::Neutral) = impl_div(Val(3), x, y)
+broadcasted(::typeof(/), x::Operand{Number}, y::Neutral) = impl_div(Val(2), x, y)
+broadcasted(::typeof(/), x::Neutral, y::Operand{Number}) = impl_div(Val(1), x, y)
 
 broadcasted(::typeof(\), x::Neutral, y::Neutral) = broadcasted(/, y, x)
 broadcasted(::typeof(\), x::Neutral, y::Operand{Number}) = broadcasted(/, y, x)