+++ test/aqua.jl
@@ -1,0 +1,2 @@
+using Aqua
+Aqua.test_all(Neutrals; stale_deps = !isdefined(Base, :get_extension))